[FEATURE] JS changes available BE variant stock
With activated stock handling for BE variants
the JavaScript which changes the prices does
also change change the available stock.

Requirements:
-This requires that the select view contains
a data atrribute 'data-available-stock' with
the stock for the BE variant.
- Furthermore some HTML has to be present which
will be used to insert the available stock.

Both changes will be made in a commit in
EXT:cart_products.

!!!
This commit changes also the buggy behaviour that
reselecting "Please choose..." from the select
field shows the price of the last chosen item.
Instead a `&nbsp;` will be inserted (to display
no price while avoiding 'jumping' layout.

Fixes: #487